+++ cli/inspire/cli/commands/account/add.py
@@ -20,8 +20,7 @@ from inspire.cli.context import Context, EXIT_VALIDATION_ERROR, pass_context
 from inspire.cli.formatters import json_formatter
 from inspire.cli.utils.errors import exit_with_error
 from inspire.cli.utils.output import emit_success
-
-DEFAULT_BASE_URL = "https://qz.sii.edu.cn"
+from inspire.config import DEFAULT_BASE_URL
 
 
 @click.command("add")